requirements.txt
PyQt5>=5.15.9
opencv-python>=4.8.0
numpy>=1.24.0
mss>=9.0.1
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
Pillow>=10.0.0
pyinstaller>=6.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
pyaudiowpatch>=0.2.12
